Fix failing test
I had missed a dtype mismatch that made the printer not work for
the `list_max` test.

Also, it's pretty clear to me at this point that I'm going to need
`ppx_deriving` in order to not get annoyed at the difficulty of
writing tests that print out actual vs expected.